move files to data-feed-service
