+++ components/data-feed-service/config/config.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	platform_config "github.com/chef/automate/lib/platform/config"
-	"github.com/chef/automate/lib/stringutils"
 	"github.com/chef/automate/lib/tls/certs"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
@@ -139,14 +138,16 @@ func Configure() (*DataFeedConfig, error) {
 		var err error
 		log.Infof("Database %s", config.PostgresConfig.Database)
 		config.PostgresConfig.ConnectionString, err = platform_config.PGURIFromEnvironment(config.PostgresConfig.Database)
-		maskedDBConnString := stringutils.MaskPGCredInURI(config.PostgresConfig.ConnectionString)
+		maskedDBConnString := MaskPGCredInURI(config.PostgresConfig.ConnectionString)
 		log.Infof("Masked Database connection string %s", maskedDBConnString)
 		if err != nil {
 			log.WithError(err).Error("Failed to get pg uri")
 			return nil, err
 		}
 	}
-	log.Debugf("DATA FEED SERVICE CONFIG: %+v", config)
+	logConfigForDebug := fmt.Sprintf("DATA FEED SERVICE CONFIG: %+v", config)
+	maskedLogConfigForDebug := MaskPGCredInURI(logConfigForDebug)
+	log.Debug(maskedLogConfigForDebug)
 	log.Debug("end config.go Configure() ->")
 	return config, nil
 }
